+++ crates/starknet-devnet-core/src/starknet/mod.rs
@@ -1167,8 +1167,8 @@ impl Starknet {
         to_block: Option<BlockId>,
         address: Option<ContractAddress>,
         keys: Option<Vec<Vec<Felt>>>,
-        skip: usize,
-        limit: Option<usize>,
+        skip: u64,
+        limit: Option<u64>,
     ) -> DevnetResult<(Vec<EmittedEvent>, bool)> {
         events::get_events(self, from_block, to_block, address, keys, skip, limit)
     }
